Moved some code from the various activity classes up to BaseActivity because I found a way to make the swipe-left and swipe-right code common.  Just created a 'currentActivity' protected field in BaseActivity.  Then in each of the subclasses set currentActivity equal to whatever class we're in.
Also fixed a bug in MainActivity.onActivityResult().  Turns out we were trying to set the text of a label on MainActivity using a Runnable.  And we were doing this right before we were switching over to LimboActivity.  So what happened - the app would crash, presumably because findViewById() was returning null since we were now on LimboActivity.
