+++ app/src/main/java/com/brentdunklau/telepatriot_android/DirectorActivity.java
@@ -2,6 +2,7 @@ package com.brentdunklau.telepatriot_android;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 
 import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.SlideIt;
 import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.SwipeAdapter;
@@ -13,55 +14,27 @@ import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android
 
 public class DirectorActivity extends BaseActivity implements SlideIt {
 
+    private final static String TAG = "DirectoryActivity";
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_director);
+        currentActivity = this.getClass();
         swipeAdapter = new SwipeAdapter(this, this);
         user = User.getInstance();
-    }
-
-
-    private Class onTheLeft() {
-        return user.activityOnTheLeft(DirectorActivity.class);
-    }
-
-
-    private Class onTheRight() {
-        return user.activityOnTheRight(DirectorActivity.class);
-    }
-
-
-    @Override
-    public void rightToLeft() {
-        Class onTheRight = onTheRight();
-        if(onTheRight != null) {
-            Intent it = new Intent(this, onTheRight());
-            startActivity(it);
-            overridePendingTransition(R.anim.slide_from_right, R.anim.slide_to_left);
-        }
-    }
-
-    @Override
-    public void leftToRight() {
-        Class onTheLeft = onTheLeft();
-        if(onTheLeft != null) {
-            Intent it = new Intent(this, onTheLeft());
-            startActivity(it);
-            overridePendingTransition(R.anim.slide_from_left, R.anim.slide_to_right);
-        }
-    }
-
-/*
-    @Override
-    protected void onPause() {
-        super.onPause();
-        Log.d("DirectorActivity", "paused");
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        Log.d("DirectorActivity", "resume");
-    }*/
+        currentActivity = this.getClass();
+        Log.d(TAG, "resume");
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        currentActivity = this.getClass();
+    }
 }
