+++ app/src/main/java/com/brentdunklau/telepatriot_android/LimboActivity.java
@@ -66,7 +66,6 @@ public class LimboActivity extends BaseActivity {
         };
 
         accountStatusEvents.setAdapter(mAdapter);
-
     }
 
     @Override
